requirements.txt
discord.py==1.7.3
discord-pretty-help==1.3.3
aiohttp==3.7.4.post0
requests==2.25.1